refactor: 나의 감정 리팩터링
