+++ app/src/main/java/com/team/recordream/presentation/record/adapter/RecordViewHolder.kt
@@ -9,8 +9,8 @@ import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.team.recordream.databinding.ItemListEmotionBinding
 import com.team.recordream.presentation.record.model.EmotionState
-import com.team.recordream.presentation.record.uistate.Emotion
-import com.team.recordream.presentation.record.uistate.Emotion.*
+import com.team.recordream.presentation.record.model.Emotion
+import com.team.recordream.presentation.record.model.Emotion.*
 
 class RecordViewHolder(
     onClick: (emotionId: Int) -> Unit,
@@ -22,7 +22,6 @@ class RecordViewHolder(
     }
 
     fun bind(emotion: EmotionState) {
-        Log.d("12312399", emotion.selected.toString())
         binding.emotion = emotion
     }
