--- app/src/main/java/com/team/recordream/presentation/record/uistate/EmotionUiState.kt
@@ -1,6 +1,0 @@
-package com.team.recordream.presentation.record.uistate
-
-data class EmotionUiState(
-    val emotion: Emotion,
-    val selected: Boolean,
-)
